requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
flask==3.1.1
sentence-transformers>=3.0.0
pypdf>=5.0.0
python-dotenv==1.1.0
langchain-pinecone>=0.2.0
langchain-community>=0.3.0
pinecone-client>=5.0.1
langchain-google-genai>=2.0.0s